+++ src/textfile_sub_source.hh
@@ -36,17 +36,16 @@
 #include "filter_observer.hh"
 #include "logfile.hh"
 #include "plain_text_source.hh"
+#include "text_link_handler.hh"
 #include "text_overlay_menu.hh"
 #include "textview_curses.hh"
 
 class textfile_sub_source
-    : public text_sub_source
+    : public text_link_handler
     , public vis_location_history
     , public text_accel_source
     , public text_anchors {
 public:
-    using file_iterator = std::deque<std::shared_ptr<logfile>>::iterator;
-
     textfile_sub_source() { this->tss_supports_filtering = true; }
 
     bool empty() const { return this->tss_files.empty(); }
@@ -81,7 +80,7 @@ public:
             return nullptr;
         }
 
-        return this->tss_files.front();
+        return this->tss_files.front().fvs_file;
     }
 
     std::string text_source_name(const textview_curses& tv) override
@@ -90,7 +89,7 @@ public:
             return "";
         }
 
-        return this->tss_files.front()->get_filename();
+        return this->tss_files.front().fvs_file->get_filename();
     }
 
     void to_front(const std::shared_ptr<logfile>& lf);
@@ -149,7 +148,7 @@ public:
     std::optional<std::string> anchor_for_row(vis_line_t vl) override;
 
     std::optional<vis_line_t> adjacent_anchor(vis_line_t vl,
-                                                 direction dir) override;
+                                              direction dir) override;
 
     std::unordered_set<std::string> get_anchors() override;
 
@@ -177,6 +176,34 @@ private:
         delete lfo;
     }
 
+    struct file_view_state {
+        explicit file_view_state(const std::shared_ptr<logfile>& f)
+            : fvs_file(f)
+        {
+        }
+
+        bool operator==(const std::shared_ptr<logfile>& lf) const
+        {
+            return this->fvs_file == lf;
+        }
+
+        void save_from(const textview_curses& tc)
+        {
+            this->fvs_top = tc.get_top();
+            this->fvs_selection = tc.get_selection();
+        }
+
+        void load_into(textview_curses& tc) const
+        {
+            tc.set_selection(this->fvs_selection);
+            tc.set_top(this->fvs_top);
+        }
+
+        std::shared_ptr<logfile> fvs_file;
+        vis_line_t fvs_top{0};
+        vis_line_t fvs_selection{0};
+    };
+
     struct rendered_file {
         time_t rf_mtime;
         file_ssize_t rf_file_size;
@@ -189,8 +216,9 @@ private:
         lnav::document::metadata ms_metadata;
     };
 
-    std::deque<std::shared_ptr<logfile>> tss_files;
-    std::deque<std::shared_ptr<logfile>> tss_hidden_files;
+    using file_iterator = std::deque<file_view_state>::iterator;
+
+    std::deque<file_view_state> tss_files;
     std::unordered_map<std::string, rendered_file> tss_rendered_files;
     std::unordered_map<std::string, metadata_state> tss_doc_metadata;
     size_t tss_line_indent_size{0};