+++ src/textview_curses.hh
@@ -509,6 +509,8 @@ public:
 
     virtual void scroll_invoked(textview_curses* tc);
 
+    virtual void text_open_href(const std::string& href) {}
+
     bool tss_supports_filtering{false};
     bool tss_apply_filters{true};
 
@@ -808,6 +810,7 @@ public:
         int64_t sti_line;
         line_range sti_range;
         std::string sti_value;
+        std::string sti_href;
     };
 
     std::optional<selected_text_info> tc_selected_text;